requirement.txt
flask
flask_cors
python-dotenv
redis
flask-redis